+++ test/pseudo_url.js
@@ -16,7 +16,7 @@ describe('Apify.PseudoUrl', () => {
         expect(purl.matches('http://www.example.com/pages/not@working')).to.be.eql(true);
     });
 
-    it('createRequest() should work', () => {
+    it('createRequest() should work with a string', () => {
         const purl = new Apify.PseudoUrl('something', { method: 'POST', userData: { foo: 'bar' } });
         const request = purl.createRequest('http://example.com');
 
@@ -24,5 +24,20 @@ describe('Apify.PseudoUrl', () => {
         expect(request.url).to.be.eql('http://example.com');
         expect(request.method).to.be.eql('POST');
         expect(request.userData).to.be.eql({ foo: 'bar' });
+    });
+
+    it('createRequest() should work with an object', () => {
+        const purl = new Apify.PseudoUrl('something', { method: 'POST', userData: { foo: 'bar' } });
+        const request = purl.createRequest({
+            url: 'http://example.com',
+            userData: {
+                bar: 'foo',
+            },
+        });
+
+        expect(request).to.be.an.instanceof(Apify.Request);
+        expect(request.url).to.be.eql('http://example.com');
+        expect(request.method).to.be.eql('POST');
+        expect(request.userData).to.be.eql({ foo: 'bar', bar: 'foo' });
     });
 });