+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 xxx
 ====================
 - Bugfix in BasicCrawler: async calls in `isFinishedFunction` were not awaited
+- Better logging of memory overload errors
 
 0.14.15 / 2019-05-31
 ====================