+++ CHANGELOG.md
@@ -13,6 +13,10 @@ NEXT
   to add requests to the queue from pure JavaScript navigations, form submissions etc.
 - The `RequestQueue.handledCount()` function has been resurrected from deprecation,
   in order to have compatible interface with `RequestList`.
+- Add `useExtendedUniqueKey` option to `Request` constructor to include `method` and `payload`
+  in the `Request`'s computed `uniqueKey`.
+- Updated Puppeteer to 1.18.0
+  
 
 0.14.15 / 2019-05-31
 ====================